Don't assume the manager and menu have the same lifetime
Summary:
The menu we retun can outlive the manager. So we can't use it or its members in
slots when connecting to signals of the menu or it's actions. This also switches
the actions' data() back to being the schemePaths instead of the rows of the
corresponding indices.

Test Plan: Open Labplot which currently triggers this crash.

Reviewers: #frameworks, broulik

Reviewed By: broulik

Subscribers: anthonyfieroni, broulik, kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D26848