Introduce the Header color set
Summary:
This introduces two concepts: a new color set for Header (or: titlebar?
open question)
and the possibility for a color set to have a different config group (a
sub group) for different states (there, for inactive)

the Header area should probably also going to slowly relace the WM
section of color themes

This is needed for implement properly T10201, this graphical union
between the titlebar and the toolbar.

There is a problem of NColorSets that has been recently added,
because it makes impossible to hard/impossible to
properly add a set in a 100% abi compatible way, strictly speaking it
should be compatible tough keeping NColorSets as the last value, is a
kindof behavior change, though *should* be acceptable (old users would
get the Header value as NColorSets but everything working never the less)

Test Plan: doing a color set with Header group, the proper values get read

Reviewers: #vdg, #plasma, cblack, ndavis

Reviewed By: #vdg, #plasma, cblack, ndavis

Subscribers: davidre, ndavis, cblack, kde-frameworks-devel

Tags: #frameworks

Maniphest Tasks: T10201

Differential Revision: https://phabricator.kde.org/D29232