Enforce linking against KF::ConfigGui
We don't actually use anything from there directly, so linking with
--as-needed will drop that dependency. We however rely on it being
loaded in order for QColor to be readable from KConfig.

As a workaround, make sure we use a symbol to force the loading. The
cleaner approach would probably to export the init symbol in KF::ConfigGui
that's used in the static linking case there.