Make KColorScheme::FocusColor QPalette::Accent
This is an alternative to
https://invent.kde.org/frameworks/kcolorscheme/-/merge_requests/15

This doesn't require new keys in the color file, but partly repurposes
Focus to be Accent. Semantically Focus is a subset of Accent, rather
the other way around, but is a way better choice than the highlight background,
as accent and highlight background need to be 2 separate and different (even if
usually related) colors.