Additional public API
Currently we can find a color scheme by name but can only get the current scheme id making it hard to get the initial index to set for example a combobox.

Add public functions to get the current scheme name and find a scheme by id so that a consumer can search either way.