Use Qt API instead of KColorSchemeWatcher
Qt has all the API we need, at least since Qt 6.10.

We cannot depend on Qt 6.10 yet though. The only thing that
KColorSchemeManager does that isn't available in Qt < 6.10
is support for detecting high-contrast mode on Windows. Since
that is only a little bit of code we can include that here