Fix colors in kdeglobals not being respected when platformtheme is not kde
Previously, applications such as Dolphin, Okular, Kate, etc. that (partially) themed their UI based on colors specified in a kdeglobals (e.g. ~/.config/kdeglobals) file could do so regardless of whether the Qt platformtheme was set to kde, so long as key ColorScheme under [UiSettings] was set to some arbitrary value (e.g. ColorScheme=*). This enabled users to properly theme such applications even when using a different platformtheme (e.g. when on a different DE or just a WM with something like qt6ct).

Due to some minor modifications made as part of framework release v6.19.0 in order to use the new id-based approach for color schemes, this no longer works and defaults are now always used instead of the values specified in kdeglobals when the platformtheme is not kde. This change seems to be unintentional. Previously, the responsible code branched based on whether kdeglobal's ColorScheme was set, with defaults being used if ColorScheme was empty/not set. In v6.19.0, it now first tries to find the QModelIndex of the scheme specified by ColorScheme and then find the scheme's id from that, then branch based on whether the scheme's id is set. If the platformtheme is not set to kde, the returned QModelIndex will be invalid and thus the scheme's id will be considered empty, regardless of the value of ColorScheme.

Simply reverting the check to look at the value of ColorScheme directly again fixes the issue and doesn't otherwise change behaviour for the new id-based approach.